% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{attach_duckdb_to_postgres}
\alias{attach_duckdb_to_postgres}
\title{Attach DuckDB to Postgres for streaming reads}
\usage{
attach_duckdb_to_postgres(con)
}
\arguments{
\item{con}{A DuckDB connection.}
}
\description{
Attach DuckDB to Postgres for streaming reads
}
