% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_indexes}
\alias{check_indexes}
\title{List indexes in Postgres}
\usage{
check_indexes(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
A tibble with schema, table, index name, and definition.
}
\description{
List indexes in Postgres
}
