% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_primary_key_exists}
\alias{check_primary_key_exists}
\title{Check if a primary key exists}
\usage{
check_primary_key_exists(
  con,
  pg_schema,
  table_name,
  column_name,
  return_constraints_table = FALSE
)
}
\arguments{
\item{con}{A DBI connection.}

\item{pg_schema}{Schema name.}

\item{table_name}{Table name.}

\item{column_name}{Column name.}

\item{return_constraints_table}{Logical, return full table if TRUE.}
}
\value{
Logical or tibble.
}
\description{
Check if a primary key exists
}
