% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_primary_keys}
\alias{check_primary_keys}
\title{List primary keys in Postgres}
\usage{
check_primary_keys(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
A tibble with schema, table, and column names.
}
\description{
List primary keys in Postgres
}
