% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{connect_to_postgres_db}
\alias{connect_to_postgres_db}
\title{Connect to a Postgres database}
\usage{
connect_to_postgres_db(
  expose_connection_to_viewer = FALSE,
  sslmode = "prefer",
  print_connection_info = FALSE
)
}
\arguments{
\item{expose_connection_to_viewer}{Logical, show in connections pane.}

\item{sslmode}{SSL mode for connection.}

\item{print_connection_info}{Logical, print connection info.}
}
\value{
A DBI connection object.
}
\description{
Reads connection parameters from environment variables.
}
