% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copy_all_tables_from_postgres_to_duckdb}
\alias{copy_all_tables_from_postgres_to_duckdb}
\title{Copy all relevant tables from Postgres to DuckDB}
\usage{
copy_all_tables_from_postgres_to_duckdb(
  con_duck,
  con_pg,
  pg_schema = "public",
  relevant_tables = c("Gitterzellen_Gemeinde_Mapping", "Gitterzellen_Einwohner_Mapping",
    "Verwaltungsgebiete_Mapping", "Gemeindegrenzen_Polygone", "Kreisgrenzen_Polygone",
    "Landesgrenzen_Polygone", "Regierungsbezirksgrenzen_Polygone",
    "Krankenhaus_Standortliste"),
  geometry_tables = c("Gemeindegrenzen_Polygone", "Kreisgrenzen_Polygone",
    "Landesgrenzen_Polygone", "Regierungsbezirksgrenzen_Polygone",
    "Krankenhaus_Standortliste"),
  geometry_column = "geometry"
)
}
\arguments{
\item{con_duck}{DuckDB connection.}

\item{con_pg}{Postgres connection.}

\item{pg_schema}{Schema name.}

\item{relevant_tables}{Character vector of table names.}

\item{geometry_tables}{Character vector of table names with geometry columns.}

\item{geometry_column}{Name of the geometry column in geometry tables.}
}
\description{
Copy all relevant tables from Postgres to DuckDB
}
