% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copy_table_from_postgres_to_duckdb}
\alias{copy_table_from_postgres_to_duckdb}
\title{Copy a table from Postgres to DuckDB}
\usage{
copy_table_from_postgres_to_duckdb(
  con_duck,
  con_pg,
  pg_schema,
  table_name,
  geometry_column = NULL
)
}
\arguments{
\item{con_duck}{DuckDB connection.}

\item{con_pg}{Postgres connection.}

\item{pg_schema}{Schema name.}

\item{table_name}{Table name.}

\item{geometry_column}{Name of geometry column to convert (NULL for non-spatial tables).}
}
\description{
Copy a table from Postgres to DuckDB
}
