% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{detach_duckdb_from_postgres}
\alias{detach_duckdb_from_postgres}
\title{Detach DuckDB from Postgres}
\usage{
detach_duckdb_from_postgres(con)
}
\arguments{
\item{con}{A DuckDB connection.}
}
\description{
Detach DuckDB from Postgres
}
