% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_strategies.R
\name{disconnect_source}
\alias{disconnect_source}
\title{Disconnect from source}
\usage{
disconnect_source(source, connections, shutdown = TRUE)
}
\arguments{
\item{source}{Source configuration (pg_source or duckdb_source)}

\item{connections}{List of connections returned by connect_to_source}

\item{shutdown}{Whether to shutdown connections (default: TRUE)}
}
\description{
Cleans up connections to the data source.
}
\keyword{internal}
