% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ensure_dir_exists}
\alias{ensure_dir_exists}
\title{Ensure a directory exists, create if not}
\usage{
ensure_dir_exists(dir_path)
}
\arguments{
\item{dir_path}{Directory path}
}
\description{
Ensure a directory exists, create if not
}
