% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{export_all_duckdb_tables_to_parquet}
\alias{export_all_duckdb_tables_to_parquet}
\title{Export all DuckDB tables to Parquet files}
\usage{
export_all_duckdb_tables_to_parquet(
  con,
  output_dir,
  large_tables_list = c("Entfernungsdaten"),
  compression = "SNAPPY"
)
}
\arguments{
\item{con}{DuckDB connection.}

\item{output_dir}{Output directory.}

\item{large_tables_list}{Vector of large table names.}

\item{compression}{Compression algorithm.}
}
\description{
Export all DuckDB tables to Parquet files
}
