% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{export_single_duckdb_table_to_parquet}
\alias{export_single_duckdb_table_to_parquet}
\title{Export a DuckDB table to Parquet}
\usage{
export_single_duckdb_table_to_parquet(
  con,
  table_name,
  output_path,
  compression = "SNAPPY"
)
}
\arguments{
\item{con}{DuckDB connection.}

\item{table_name}{Table name.}

\item{output_path}{Output file path.}

\item{compression}{Compression algorithm.}
}
\description{
Export a DuckDB table to Parquet
}
