% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_tables_and_sizes}
\alias{list_tables_and_sizes}
\title{List tables and their sizes in Postgres}
\usage{
list_tables_and_sizes(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
A tibble with table names and sizes.
}
\description{
List tables and their sizes in Postgres
}
