% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_temp_tables}
\alias{list_temp_tables}
\title{List temporary tables in Postgres}
\usage{
list_temp_tables(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
A tibble with schema, table, column, and type.
}
\description{
List temporary tables in Postgres
}
