% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_geoparquet}
\alias{read_geoparquet}
\title{Read a GeoParquet file into an sf object}
\usage{
read_geoparquet(path)
}
\arguments{
\item{path}{Path to the GeoParquet file.}
}
\value{
An sf object with geometry column.
}
\description{
Convenience wrapper for reading GeoParquet files exported by this package.
Requires the sf and arrow packages to be installed.
}
\examples{
\dontrun{
boundaries <- read_geoparquet("output/Gemeindegrenzen_Polygone.parquet")
plot(boundaries["Gemeindeschluessel"])
}

}
