% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_index}
\alias{remove_index}
\title{Remove an index from Postgres}
\usage{
remove_index(con, pg_schema, index_name)
}
\arguments{
\item{con}{A DBI connection.}

\item{pg_schema}{Schema name.}

\item{index_name}{Index name.}
}
\description{
Remove an index from Postgres
}
