% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_table}
\alias{remove_table}
\title{Remove a table from Postgres}
\usage{
remove_table(con, pg_schema, table_name)
}
\arguments{
\item{con}{A DBI connection.}

\item{pg_schema}{Schema name.}

\item{table_name}{Table name.}
}
\description{
Remove a table from Postgres
}
