% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{require_file_exists}
\alias{require_file_exists}
\title{Require a file exists, stop with error if not}
\usage{
require_file_exists(file_path)
}
\arguments{
\item{file_path}{File path}
}
\description{
Require a file exists, stop with error if not
}
