% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_index}
\alias{set_index}
\title{Add an index to a table}
\usage{
set_index(con, pg_schema, tbl, col, idx_name)
}
\arguments{
\item{con}{A DBI connection.}

\item{pg_schema}{Schema name.}

\item{tbl}{Table name.}

\item{col}{Column name.}

\item{idx_name}{Index name.}
}
\description{
Add an index to a table
}
