% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_primary_key}
\alias{set_primary_key}
\title{Add a primary key to a table}
\usage{
set_primary_key(con, table_name, columns)
}
\arguments{
\item{con}{A DBI connection.}

\item{table_name}{Table name.}

\item{columns}{Character vector of column names.}
}
\description{
Add a primary key to a table
}
