% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SMCReichbR-package}
\alias{SMCReichbR}
\alias{SMCReichbR-package}
\title{SMCReichbR: SMC Tools for Krankenhaus Erreichbarkeitsanalysen}
\description{
Provides tools for accessing the Science Media Center (SMC) hospital driving time database and performing hospital accessibility analyses (Erreichbarkeitsanalysen). Functions are included to connect to PostgreSQL and DuckDB databases, load and transform geospatial data, apply geographic filters, and export summaries for downstream reporting and visualization.
}
\author{
\strong{Maintainer}: Simon Essink \email{simon.essink@sciencemediacenter.de} (\href{https://orcid.org/0000-0003-1390-3885}{ORCID})

}
\keyword{internal}
