% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{connect_to_duckdb_db}
\alias{connect_to_duckdb_db}
\title{Connect to a DuckDB database}
\usage{
connect_to_duckdb_db(
  db_path,
  expose_connection_to_viewer = FALSE,
  print_connection_info = FALSE
)
}
\arguments{
\item{db_path}{Path to DuckDB file.}

\item{expose_connection_to_viewer}{Logical, show in connections pane.}

\item{print_connection_info}{Logical, print connection info.}
}
\value{
A DBI connection object.
}
\description{
Connect to a DuckDB database
}
