% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ReichbR-package}
\alias{ReichbR}
\alias{ReichbR-package}
\title{ReichbR: SMC Tools for Krankenhaus Erreichbarkeitsanalysen}
\description{
Provides helpers for accessing the SMC hospital driving time database and for running Erreichbarkeitsanalysen. The package bundles utilities to load and transform the source data, to filter spatial selections, and to export summaries for downstream reporting.
}
\author{
\strong{Maintainer}: Simon Essink \email{simon.essink@sciencemediacenter.de} (\href{https://orcid.org/0000-0003-1390-3885}{ORCID})

}
\keyword{internal}
